Changed save data
Changed save data generated by the game editor to be XML rather than a dump of the binary data of classes. This is so that during development and adding more variables ect to classes is easier to manage with current test maps. The functionality to toggle between save/load formats should be added in the future as binary data is faster to load also with a smaller filesize as long as the class structure stays the same. A toggle for both save and load format will also make it possible to change the format if a class has an unplanned change in its structure.